Alertest/pyproject.toml
[tool.poetry]
name = "alertest"
version = "0.1.0"
description = ""
authors = ["''"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.13.0"
aiohttp = { extras = ["speedups"], version = "3.11.4" }
loguru = "0.7.2"
nats-py = "2.9.0"
orjson = "3.10.11"
python-decouple = "3.8"
setuptools = "75.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
